Remove oc adm create-node-config
