+++ hack/lib/start.sh
@@ -37,7 +37,6 @@ function os::start::configure_server() {
 	current_user="$( id -u )"
 
 	os::start::internal::create_master_certs     "${version}"
-	os::start::internal::configure_node          "${version}"
 	os::start::internal::configure_master        "${version}"
 
 	# fix up owner after creating initial config
@@ -71,37 +70,6 @@ function os::start::internal::create_master_certs() {
 	                        --public-master="${API_SCHEME}://${PUBLIC_MASTER_HOST}:${API_PORT}"
 }
 readonly -f os::start::internal::create_master_certs
-
-# os::start::internal::configure_node creates a node configuration
-#
-# Globals:
-#  - NODE_CONFIG_DIR
-#  - KUBELET_SCHEME
-#  - KUBELET_BIND_HOST
-#  - KUBELET_PORT
-#  - KUBELET_HOST
-#  - MASTER_ADDR
-#  - MASTER_CONFIG_DIR
-# Arguments:
-#  1 - alternate version for the config
-function os::start::internal::configure_node() {
-	local version="${1:-}"
-	local openshift_volumes=( "${MASTER_CONFIG_DIR}" "${NODE_CONFIG_DIR}" )
-
-	os::log::debug "Creating node configuration for the OpenShift server"
-	oc adm create-node-config                                          \
-	                        --node-dir="${NODE_CONFIG_DIR}"                                   \
-	                        --node="${KUBELET_HOST}"                                          \
-	                        --hostnames="${KUBELET_HOST}"                                     \
-	                        --master="${MASTER_ADDR}"                                         \
-	                        --signer-cert="${MASTER_CONFIG_DIR}/ca.crt"                       \
-	                        --signer-key="${MASTER_CONFIG_DIR}/ca.key"                        \
-	                        --signer-serial="${MASTER_CONFIG_DIR}/ca.serial.txt"              \
-	                        --certificate-authority="${MASTER_CONFIG_DIR}/ca.crt"             \
-	                        --node-client-certificate-authority="${MASTER_CONFIG_DIR}/ca.crt" \
-	                        --listen="${KUBELET_SCHEME}://${KUBELET_BIND_HOST}:${KUBELET_PORT}"
-}
-readonly -f os::start::internal::configure_node
 
 # os::start::internal::configure_master creates the configuration for the OpenShift master
 #
